refactoring: merge printer with shortened names
